+++ packages/media-query-list-parser/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased (patch)
 
+- Fixed: avoid exponential parse times for deeply nested media conditions.
 - Updated [`@csstools/css-tokenizer`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-tokenizer) to [`4.0.1`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-tokenizer/CHANGELOG.md#401) (patch)
 
 ### 5.0.0